Add import/export DB feature
Add possibility to import/export databases to facilitate moving data
between clients (as the Web GUI does not have a server backend).
Currently native Dexie feature is used for full DB import and export
that allows cloning of the DB to an other device. The exported file is
in JSON format which allows editing and manipulating if needed.