Update animation due to v20 changes
Angular v20 update disabled animations when prefers-reduced-motion is
set to `reduced` due to disabled animation e.g. in windows. This commit
overrides the `MediaMatcher` to always return no-preference, allowing
animations to work as expected. Furthermore, it adds back mat-error
animations that seems to be missing in the update.

Furthermore as v20 deprecates the animation API this commit refactors
the code to use CSS based animation.