Convert deviceInfo to signal
This commit refactors ErgGenericDataService to expose deviceInfo as a
Signal derived from connectionStatus$ and the readDeviceInfo method
(which is now private). Further updates OpenSettingsButtonComponent to
use the new API as well as improve opening behaviour by showing the main
spinner only when connections status is `connecting` and waiting for the
connection process to complete before opening the settings dialog. Also
add comprehensive unit tests for OpenSettingsButtonComponent.